+++ commons/util/src/main/java/net/automatalib/common/util/nid/DynamicList.java
@@ -140,7 +140,7 @@ public class DynamicList<T extends MutableNumericID> extends AbstractList<T> {
 
     @Override
     public Iterator<T> iterator() {
-        return new Iterator<T>() {
+        return new Iterator<>() {
 
             int index;